chore: move test only packages behind a build flag
